增加mkt.campaign.progressstatus.list.get接口

+++ marketing-tool-service/src/main/java/cn/rongcapital/mkt/service/CampaignProgressStatusListService.java
@@ -18,6 +18,6 @@ public interface CampaignProgressStatusListService {
 	 * @param publishStatus campaignName index size
 	 * @return 
 	 */
-	BaseOutput campaignProgressStatusList(Integer publishStatus,
+	BaseOutput campaignProgressStatusList(Byte publishStatus,
 			String campaignName, Integer index, Integer size);
 }
